Fix routeCheck failures for Arista-7050CX3-32S-C28S16 and Arista-7050CX3-32C-C28S16 (#23895)
<!--
 Please make sure you've read and understood our contributing guidelines:
 https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

 failure_prs.log skip_prs.log Make sure all your commits include a signature generated with `git commit -s` **

 If this is a bug fix, make sure your description includes "fixes #xxxx", or
 "closes #xxxx" or "resolves #xxxx"

 Please provide the following information:
-->

#### Why I did it

This is the issue tracked in https://github.com/aristanetworks/sonic-qual.msft/issues/701

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it

This PR fixes this issue by fixing the configuration of `Arista-7050CX3-32S-C28S16` and `Arista-7050CX3-32C-C28S16`, based on the configuration of other HWSKUs for the same platforms.

#### How to verify it

Verified the `routeCheck` is OK with this change, e.g.
```
Monit 5.20.0 uptime: 3h 59m

Program 'routeCheck'
 status Status ok
 monitoring status Waiting
 monitoring mode active
 on reboot start
 last exit value 0
 last output -
 data collected Thu, 28 Aug 2025 16:28:50
```

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ ] 202205
- [ ] 202211
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [x] 202505

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [x] 202505

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

<!--
 Ensure to add label/tag for the feature raised. example - PR#2174 under sonic-utilities repo. where, Generic Config and Update feature has been labelled as GCU.
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->

#### A picture of a cute animal (not mandatory but encouraged)