[202505][Moby] Enable xcvrd SFF manager (#24262)
Why I did it
High power mode is not correctly enabled for non CMIS QSFP modules, and we plan to use an AOI modules in the Moby QSFP port that requires enabling high power mode in 202505.

How I did it
This change enables SFF manager so that high power mode can be correctly set for the API opitcs (and other affected modules). Note that we want to limit the impact in 202505 so we only enable it on Moby.

How to verify it
Confirm the same AOI optics, which couldn't link up, can work with this change
Run sonicmgmt tests to make sure no regressions are introduced
Which release branch to backport (provide reason below if selected)
This change only targets 202505